StringParsingBuffer's operator+=() should not take in a signed integer type https://bugs.webkit.org/show_bug.cgi?id=310521
Reviewed by Darin Adler.

StringParsingBuffer's operator+=() should not take in a signed integer
type. It currently takes in a signed integer (int) and then relies on
`advanceBy()` internally which takes in an unsigned integer (size_t).

* Source/WTF/wtf/text/StringParsingBuffer.h:

Canonical link: https://commits.webkit.org/309754@main

+++ Source/WTF/wtf/text/StringParsingBuffer.h
@@ -121,7 +121,7 @@ public:
         return result;
     }
 
-    constexpr StringParsingBuffer& operator+=(int places)
+    constexpr StringParsingBuffer& operator+=(size_t places)
     {
         advanceBy(places);
         return *this;
